feat(storage): support delete source objects on compose
Updates compose file sample to support deleting source objects optionally. Fixes b/441557254

[Generated-by: AI]

+++ storage/test/ObjectsTest.php
@@ -260,7 +260,10 @@ EOF;
         return [[true], [false]];
     }
 
-    public function testCompose()
+    /**
+     * @dataProvider provideCompose
+     */
+    public function testCompose(bool $deleteSourceObjects)
     {
         $bucket = self::$storage->bucket(self::$bucketName);
         $object1Name = uniqid('compose-object1-');
@@ -269,26 +272,48 @@ EOF;
         $bucket->upload('content', ['name' => $object2Name]);
 
         $targetName = uniqid('compose-object-target-');
-        $output = self::runFunctionSnippet('compose_file', [
+        
+        $args = [
             self::$bucketName,
             $object1Name,
             $object2Name,
             $targetName,
-        ]);
+        ];
+        
+        if ($deleteSourceObjects) {
+            $args[] = true;
+        }
+        
+        $output = self::runFunctionSnippet('compose_file', $args);
 
-        $this->assertEquals(
-            sprintf(
-                'New composite object %s was created by combining %s and %s',
-                $targetName,
-                $object1Name,
-                $object2Name
-            ),
-            $output
+        $expectedOutput = sprintf(
+            'New composite object %s was created by combining %s and %s',
+            $targetName,
+            $object1Name,
+            $object2Name
         );
 
-        $bucket->object($object1Name)->delete();
-        $bucket->object($object2Name)->delete();
+        if ($deleteSourceObjects) {
+            $expectedOutput .= ' and source objects were deleted';
+        }
+        $expectedOutput .= PHP_EOL;
+
+        $this->assertEquals($expectedOutput, $output);
+
+        $this->assertEquals(!$deleteSourceObjects, $bucket->object($object1Name)->exists());
+        $this->assertEquals(!$deleteSourceObjects, $bucket->object($object2Name)->exists());
+
+        if (!$deleteSourceObjects) {
+            $bucket->object($object1Name)->delete();
+            $bucket->object($object2Name)->delete();
+        }
+
         $bucket->object($targetName)->delete();
+    }
+
+    public function provideCompose()
+    {
+        return [[true], [false]];
     }
 
     public function testUploadAndDownloadObjectFromMemory()
